cleaned up files to start work
