+++ src/App.css
@@ -44,10 +44,6 @@ body {
   margin-bottom: 60px;
 }
 
-header {
-  /* margin-bottom: 60px; */
-}
-
 nav {
   /* background-color: var(--sub2); */
   background-color: var(--main1);
@@ -69,10 +65,6 @@ nav a {
   text-transform: lowercase;
   font-weight: 700;
   letter-spacing: 1px;
-}
-
-.nav-logo {
-  /* border: 2px solid red; */
 }
 
 .nav-logo:hover {
@@ -105,11 +97,6 @@ nav div {
 .top-nav-left {
   display: flex;
   justify-content: space-between;
-  /* width: 33%; */
-  /* border: 1px solid yellow; */
-}
-
-.top-left-right {
   /* width: 33%; */
   /* border: 1px solid yellow; */
 }
@@ -152,7 +139,7 @@ form {
   display: flex;
   flex-direction: column;
   align-items: center;
-  margin-bottom: ;
+  margin-bottom: 10px;
 }
 
 .form-page input {
